+++ tests/testthat/test_checkClass.R
@@ -23,10 +23,10 @@ test_that("checkClass", {
 
   foo = 1
   class(foo) = c("a", "b")
-  expect_error(assertClass(foo, "c"), "Must have class 'c', but has classes 'a','b'")
+  expect_error(assertClass(foo, "c"), "class 'c', but has classes 'a','b'")
   expect_error(assertClass(foo, "b", ordered=TRUE), "Must have class 'b' in position 1, but has classes 'a','b'")
 
   foo = 1
   class(foo) = "a"
-  expect_error(assertClass(foo, "c"), "Must have class 'c', but has class 'a'")
+  expect_error(assertClass(foo, "c"), "class 'c', but has class 'a'")
 })